requirements.txt
audioread==3.0.1
boto3==1.34.34
Flask==3.0.3
numpy==1.24.3
pandas==2.0.1
parselmouth==1.1.1
python-dotenv==1.0.1
scipy==1.14.1
